feat: 관광시설 리스트 리턴 api에 찜 항목 추가

+++ backend/src/main/java/com/triply/barrierfreetrip/review/dto/ReviewListDto.java
@@ -11,6 +11,6 @@ import lombok.Setter;
 @NoArgsConstructor
 public class ReviewListDto {
     private String nickname;
-    private double rating;
+    private long rating;
     private String content;
 }
